refactor : Application 구분
