+++ account/forms.py
@@ -2,10 +2,12 @@ from django import forms
 from django.contrib.auth.forms import UserCreationForm
 from django.contrib.auth.models import User
 
+
 class CustomUserCreationForm(UserCreationForm):
     class Meta:
         model = User
         fields = ['username', 'password1', 'password2']
+
 
 class CSVUploadForm(forms.Form):
     csv_file = forms.FileField()
