+++ account/templates/admin/upload_success.html
