+++ config/urls.py
@@ -1,28 +1,16 @@
-"""
-URL configuration for config project.
-
-The `urlpatterns` list routes URLs to views. For more information please see:
-    https://docs.djangoproject.com/en/4.2/topics/http/urls/
-Examples:
-Function views
-    1. Add an import:  from my_app import views
-    2. Add a URL to urlpatterns:  path('', views.home, name='home')
-Class-based views
-    1. Add an import:  from other_app.views import Home
-    2. Add a URL to urlpatterns:  path('', Home.as_view(), name='home')
-Including another URLconf   
-    1. Import the include() function: from django.urls import include, path
-    2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
-"""
 from django.contrib import admin
 from django.urls import path, include
-from courses import views
+import account
 
 urlpatterns = [
     path('admin/', admin.site.urls),
-    path('signup', views.signup_view, name='signup'),
-    path('login', views.login_view, name='login'),
-    path('logout', views.logout_view, name='logout'),
     path('courses/', include('courses.urls')),
-    path('get-csrf-token/', views.get_csrf_token, name='get_csrf_token'),
+    path('series/', include('series.urls')),
+    path('account/', include('account.urls')),
+
+    # 리팩토링
+    path('signup', account.views.signup_view, name='signup'),
+    path('login', account.views.login_view, name='login'),
+    path('logout', account.views.logout_view, name='logout'),
+    path('get-csrf-token/', account.views.get_csrf_token, name='get_csrf_token'),
 ]
