+++ courses/templates/courses/detail_test.html
@@ -1,0 +1,145 @@
+{% extends 'base_index.html' %}
+{% load static %}
+
+{% block content %}
+<div class="ms-4">
+    <nav aria-label="breadcrumb">
+      <ol class="breadcrumb bg-transparent mb-0 pb-3 pt-1 px-0 me-sm-6 me-5">
+          <h4>Tags</h4>
+          {% comment %} <li class="breadcrumb-item text-sm text-white active" aria-current="page">Python</li> {% endcomment %}
+      </ol>
+      {% comment %} <h6 class="font-weight-bolder text-white mb-0">Dashboard</h6> {% endcomment %}
+      <div class="btn-group" role="group" >
+        <button type="button" class="btn btn-primary rounded-pill me-2 btn-sm">python</button>
+        <button type="button" class="btn btn-primary rounded-pill me-2 btn-sm">django</button>
+        <button type="button" class="btn btn-primary rounded-pill btn-sm">easy</button>
+      </div>
+      </nav>
+    
+  </div>
+    <div class="container-fluid py-4">
+        <div class="row">
+            <h2>인생처럼 코딩도 실전! 👊</h2>
+        </div>
+        <div class="row" style="padding:10px">
+        
+        <div class="col-xl-2 col-sm-1-2 mb-xl-0 mb-4">
+            <div class="card">
+            <div class="card-body p-3">
+                <div class="row">
+                    <img class="border-radius-lg" src="{% static 'img/coverImage.png' %}">
+                    <div class="numbers">
+                    <p class="text-sm mb-0 text-uppercase font-weight-bold">구름</p>
+                    <h6 class="font-weight-bolder">
+                        <a href="#">MERN STACK 커뮤니티 : 시작부터 배포까지 알려주는 React</a>
+                    </h6>
+                    <p class="mb-0">
+                        <span class="text-success text-sm font-weight-bolder">+55%</span>
+                        ₩ 38,500
+                    </p>
+                    </div>
+                </div>
+            </div>
+            </div>
+        </div>
+        <div class="col-xl-2 col-sm-1-2 mb-xl-0 mb-4">
+            <div class="card">
+            <div class="card-body p-3">
+                <div class="row">
+                    <img class="border-radius-lg" src="{% static 'img/coverImage.png' %}">
+                    <div class="numbers">
+                    <p class="text-sm mb-0 text-uppercase font-weight-bold">구름</p>
+                    <h6 class="font-weight-bolder">
+                        <a href="#">MERN STACK 커뮤니티 : 시작부터 배포까지 알려주는 React</a>
+                    </h6>
+                    <p class="mb-0">
+                        <span class="text-success text-sm font-weight-bolder">+55%</span>
+                        ₩ 38,500
+                    </p>
+                    </div>
+                </div>
+            </div>
+            </div>
+        </div>
+        <div class="col-xl-2 col-sm-1-2 mb-xl-0 mb-4">
+            <div class="card">
+            <div class="card-body p-3">
+                <div class="row">
+                    <img class="border-radius-lg" src="{% static 'img/coverImage.png' %}">
+                    <div class="numbers">
+                    <p class="text-sm mb-0 text-uppercase font-weight-bold">구름</p>
+                    <h6 class="font-weight-bolder">
+                        <a href="#">MERN STACK 커뮤니티 : 시작부터 배포까지 알려주는 React</a>
+                    </h6>
+                    <p class="mb-0">
+                        <span class="text-success text-sm font-weight-bolder">+55%</span>
+                        ₩ 38,500
+                    </p>
+                    </div>
+                </div>
+            </div>
+            </div>
+        </div>
+        <div class="col-xl-2 col-sm-1-2 mb-xl-0 mb-4">
+            <div class="card">
+            <div class="card-body p-3">
+                <div class="row">
+                    <img class="border-radius-lg" src="{% static 'img/coverImage.png' %}">
+                    <div class="numbers">
+                    <p class="text-sm mb-0 text-uppercase font-weight-bold">구름</p>
+                    <h6 class="font-weight-bolder">
+                        <a href="#">MERN STACK 커뮤니티 : 시작부터 배포까지 알려주는 React</a>
+                    </h6>
+                    <p class="mb-0">
+                        <span class="text-success text-sm font-weight-bolder">+55%</span>
+                        ₩ 38,500
+                    </p>
+                    </div>
+                </div>
+            </div>
+            </div>
+        </div>
+        <div class="col-xl-2 col-sm-1-2 mb-xl-0 mb-4">
+            <div class="card">
+            <div class="card-body p-3">
+                <div class="row">
+                    <img class="border-radius-lg" src="{% static 'img/coverImage.png' %}">
+                    <div class="numbers">
+                    <p class="text-sm mb-0 text-uppercase font-weight-bold">구름</p>
+                    <h6 class="font-weight-bolder">
+                        <a href="#">MERN STACK 커뮤니티 : 시작부터 배포까지 알려주는 React</a>
+                    </h6>
+                    <p class="mb-0">
+                        <span class="text-success text-sm font-weight-bolder">+55%</span>
+                        ₩ 38,500
+                    </p>
+                    </div>
+                </div>
+            </div>
+            </div>
+        </div>
+        <div class="col-xl-2 col-sm-1-2 mb-xl-0 mb-4">
+            <div class="card">
+            <div class="card-body p-3">
+                <div class="row">
+                    <img class="border-radius-lg" src="{% static 'img/coverImage.png' %}">
+                    <div class="numbers">
+                    <p class="text-sm mb-0 text-uppercase font-weight-bold">구름</p>
+                    <h6 class="font-weight-bolder">
+                        <a href="#">MERN STACK 커뮤니티 : 시작부터 배포까지 알려주는 React</a>
+                    </h6>
+                    <p class="mb-0">
+                        <span class="text-success text-sm font-weight-bolder">+55%</span>
+                        ₩ 38,500
+                    </p>
+                    </div>
+                </div>
+            </div>
+            </div>
+        </div>
+        
+        
+        </div>
+    </div>
+
+{% endblock %}
