+++ series/urls.py
@@ -1,0 +1,9 @@
+from django.contrib import admin
+from django.urls import path, include
+from . import views
+
+app_name = 'series'
+
+urlpatterns = [
+    path('', views.index, name='index'),
+]
