+++ series/views.py
@@ -1,0 +1,7 @@
+from django.shortcuts import render
+
+# Create your views here.
+
+
+def index(request):
+    return render(request, 'series/index.html')
